fix: partially fixing the rendering issue for the EvidenceDataObject
Potential fix documented in the previous commit
