+++ src/frss-extension/elements/PotentialEvidence/index.ts
@@ -1,16 +1,16 @@
+// types
 import Element from '../../types';
+
+// element module
 import controls from './controls';
 import definition from './definition';
-// import icon from './assets/x.png';
 import properties from './properties';
-import render from './rendererEntry';
 
 const potentialEvidence: Element = {
   controls,
   definition,
   // icon,
   properties,
-  render,
 };
 
 /**
